+++ libs/ledger/ledger-consensual-data/src/main/java/net/corda/ledger/consensual/data/transaction/verifier/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.ledger.consensual.data.transaction.verifier;
+
+import org.osgi.annotation.bundle.Export;
