+++ libs/ledger/ledger-utxo-data/src/main/kotlin/net/corda/ledger/utxo/data/transaction/WrappedUtxoWireTransaction.kt
@@ -4,6 +4,7 @@ import net.corda.ledger.common.data.transaction.WireTransaction
 import net.corda.ledger.utxo.data.state.StateAndRefImpl
 import net.corda.ledger.utxo.data.state.TransactionStateImpl
 import net.corda.ledger.utxo.data.state.getEncumbranceGroup
+import net.corda.ledger.utxo.data.transaction.verifier.verifyMetadata
 import net.corda.v5.application.serialization.SerializationService
 import net.corda.v5.application.serialization.deserialize
 import net.corda.v5.crypto.SecureHash
@@ -28,6 +29,7 @@ class WrappedUtxoWireTransaction(
     }
 
     init{
+        verifyMetadata(wireTransaction.metadata)
         check(wireTransaction.componentGroupLists[UtxoComponentGroup.OUTPUTS.ordinal].size ==
                 wireTransaction.componentGroupLists[UtxoComponentGroup.OUTPUTS_INFO.ordinal].size
         ) {
